# Improve mesh loading