… src/slicer.cpp