This is a work in progress of a re-implementation of vertices merging during mesh loading, made to be more precise and more efficient. A 3DBenchy no longer triggers the warning `Mesh has overlapping faces!`.

**Background:** `Mesh::addFace` lockups vertices in a spatial hash map. This recovers the connectivity while accommodating for inaccuracies in STLs. Vertices closer than a 0.03 mm are merged together.

Previously, a vertex was inserted into a 0.03 mm wide cubic bucket. During lookup, a single bucket is searched for a neighboring vertex. Vertices close the the side of one cube will miss merge candidates from neighboring cubes. This could lead to unexpected behavior, and exacerbate the order sensitivity of the greedy merge algorithm.

This implementation uses half-open 0.064mm wide cubic buckets for merging vertices closer than 0.032 mm. It is optimized for queries: on insertion, a vertex is inserted in all of the 8 buckets overlapping with the 0.032mm radius sphere centered on the vertex. A query gets all candidates vertices from a single bucket.

For this purpose I wrote a specialized hash map. Iterating over a bucket is accomplished by iteratively probing the hash map. This gets some collisions during lookup, but they are filtered by the distance checks anyway. Using `std::unordered_map`, as in the current implementation, requires an indirection to dynamically sized buckets and can't combine collision checks with distance checks.

The hash map implementation is minimalist but supports rehashing when the number of vertices can't be estimated beforehand. Some cache locality is recovered from the inherent spatial locality of meshes: 
 * Point coordinates are hashed to their [Morton codes](https://en.wikipedia.org/wiki/Z-order_curve),
 * The probing strategy is copied from [Python's dict](https://github.com/python/cpython/blob/master/Objects/dictnotes.txt#L120),
 * Load factor is 2/3, growth factor is 4. Map sizes are powers of two, starting from 2^10.

This was a fun exercise, but I'd like to know if there is an interest in this before polishing it.
I'm in the process of benchmarking, and experimenting with different strategies and constants. (I will update with the results). But overall, the one above seems to perform the best for it's simplicity.