@Ghostkeeper 2021-12-27
> Thus, maybe the implementation should only perform strict merging, and then rely on the stitching of 2D polygons to close malformed paths. This would give another ~8x improvement in memory usage and CPU time to the `HasMap3D` implementation.

I agree with this. Since we're using it for debugging, it is nice if the mesh loading in CuraEngine works the same way as the mesh loading in Cura's front-end. This uses numpy-stl which does no deduplication. Improved performance and simplicity are bonus then.