@Piezoid 2021-12-28
I've benchmarked the **mesh loading code** with a large STL (383M binary STL, 5.4M vertices and 8M faces). Mesh is loaded through libArcus. To skip the rest of the slicing process, a `exit(3)` is added at the end of `ArcusCommunication::Private::readMeshGroupMessage`. 
For each test `/usr/bin/time -v ./CuraEngine connect 127.0.0.1:49675` is ran three times, taking the run with the best wall-clock time.

`master` ran in **8.35s** wall-clock (7s user, 1.1s system) with a maximum RSS of **1.21Go** and 755K page faults.

The present code in this PR, with vertices merging, took 5.84s (4.53s user, 0.91s system) with a maximum RSS of 1.09Go and 606K page faults. This is a 30% reduction of run time compared to master.

Removing the merging of close vertices, the run time drops to 4.24s (3.29s user, 0.77s system) with a maximum RSS of 1.1Go and 512K page faults. This adds a reduction of 27%, (49% compared to `master`)

I noticed that the maximum RSS was not affected by skipping deduplication. This is because the peak is reached during the `shrink_to_fit()` call on vertices and faces vectors, which double the memory occupation during the copy.
I added guards that only performs shrinking when more than 1/3 of the space is wasted by over-allocations. No shrinking happends when using libArcus which provide the number of faces beforehand.
This lowers the runtime to **3.96s** (3.20s user, 0.47s system) with a peak RSS of **824Mo** and 388K page faults. The size of allocations inside `Mesh` are comparable to those happening inside libArcus/Protobuf.

This was best improvement I could get by tuning this change set: **53%** reduction in mesh loading time compared to `master`.

The tests above were performed using `boost::container::small_vector`
to store the connected faces ids in instances of `MeshVertex`.
Using `std::vector` instead increases the runtime to 5.02s (+27%, 4.20s user, 0.55s system) with a peak RSS of 886Mo and 542K page faults. Memory usage and page faults increase a bit from fragmentation.


Running the **full slicing** process on the test STL with the fdmprinter's Normal - 0.15 mm profile, took 14.9s (137.6s user, 2.17s system) with a peak RSS of 1.4Go and 844K page faults, resulting in 292 layers. This is with the optimized mesh loading and `small_vector`.
Doing the same with `master` took 20.8s (143.2s user, 2.35s system) with a peak RSS of 1.93Go and 1.53M page faults.
Again, I took the best of 3 runs, but this time the results were much more volatile because of the multithreaded scheduling (on a old NUMA machine, 2x 6C12T). Still, I wasn't expecting a 27% improvement, something could be off...

Here's a screenshot of what it went through :sweat_smile:  
![image](https://user-images.githubusercontent.com/3748582/147592605-f4d7ae99-6121-4578-a740-de14601f0385.png)